refactor: move duplicate code to test/utls
